event click functioning, updates repair and event schemas
