+++ client/src/components/private/RepairForm.js
@@ -10,287 +10,17 @@ import {
   Label,
   Input,
 } from "reactstrap";
-import FullCalendar from "@fullcalendar/react";
-import dayGridPlugin from "@fullcalendar/daygrid";
-import interactionPlugin from "@fullcalendar/interaction";
-import rrulePlugin from "@fullcalendar/rrule";
-import moment from "moment";
-import API from "../../utils/API";
-import "./main.scss";
+
 
 export default class Calendar extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.calendarComponentRef = React.createRef();
-    this.state = {
-      events: [],
-      userId: this.props.userId,
-      repairId: "",
-      //   search: "",
-      repairType: "Repair",
-      title: "",
-      recurrencePeriod: "never",
-      repeatInterval: 1,
-      repeatDayOfWeek: "",
-      startDate: "",
-      recurrenceStartDate: "",
-      recurrenceEndDate: "",
-      startTime: "",
-      endTime: "",
-      cost: 0,
-      priority: "low",
-      status: "Thinking about it!",
-      isVendor: false,
-      vendor: "",
-      notes: "",
-      editable: true,
-
-      modal: false,
-      // detailModal: false,
-    };
-
-    this.handleFormSubmit = this.handleFormSubmit.bind(this);
-    this.handleInputChange = this.handleInputChange.bind(this);
-  }
-
-  componentDidMount() {
-    this.loadEvents();
-  }
-
-  loadEvents = () => {
-    API.getEvents(this.props.userId)
-      .then(response => {
-            console.log("On Load from getEvents: " + JSON.stringify(response.data));
-        this.setState({
-          events: response.data,
-        });
-      })
-      .catch(err => console.log(err));
-  };
-
-  toggle = () => {
-    this.setState(prevState => ({
-      modal: !prevState.modal,
-    }));
-  };
-
-  // Want to create modal that shows event or repair details.
-  /* handleDateClick = (dateObj) => {
-        const {date} = dateObj;
-        const newEvents = this.state.events.filter (event => event.start === date);
-        console.log("this.state.events: " + this.state.events);
-        console.log("Date: " + date);
-        console.log(newEvents);
-    } */
-
-  handleNewEvent = arg => {
-    this.setState(prevState => ({
-      modal: !prevState.modal,
-    }));
-  };
-
-  handleCheck = () => {
-    this.setState({ isVendor: !this.state.isVendor });
-  };
-
-  handleInputChange = event => {
-    // Getting the value and name of the input which triggered the change
-    const { name, value } = event.target;
-    // console.log(event.target);
-    // Updating the input's state
-    this.setState({
-      [name]: value,
-    });
-  };
-
-  handleFormSubmit = event => {
-    // Preventing the default behavior of the form submit (which is to refresh the page)
-    event.preventDefault();
-
-    let newEvent;
-
-    let newRepair = {
-      userId: this.state.userId,
-      repairType: this.state.repairType,
-      title: this.state.title,
-      cost: this.state.cost,
-      priority: this.state.priority,
-      status: this.state.status,
-      recurrencePeriod: this.state.recurrencePeriod,
-      vendor: this.state.vendor,
-      notes: this.state.notes,
-      startTime: this.state.startTime,
-      endTime: this.state.endTime,
-    };
-
-    if (this.state.recurrencePeriod !== "never") {
-      let momentStart =
-        this.state.recurrenceStartDate + " " + this.state.startTime;
-      let momentEnd = this.state.recurrenceStartDate + " " + this.state.endTime;
-      let duration = moment
-        .duration(
-          moment(momentEnd, "YYYY/MM/DD HH:mm").diff(
-            moment(momentStart, "YYYY/MM/DD HH:mm")
-          )
-        )
-        .asHours();
-
-      console.log("duration: " + duration);
-
-      let parsedRecurStart =
-        this.state.recurrenceStartDate + "T" + this.state.startTime + ":00";
-      //    let parsedRecurEnd = this.state.recurrenceEndDate + "T" + this.state.endTime + ":00";
-
-      newRepair.recurrenceStartDate = parsedRecurStart;
-
-      newRepair.repeatDayOfWeek = this.state.repeatDayOfWeek;
-      newRepair.recurrenceEndDate = this.state.recurrenceEndDate;
-      newRepair.duration = duration;
-
-      if (this.state.recurrencePeriod === "daily") {
-        this.setState({
-          repeatInterval: 1,
-        });
-
-        newRepair.repeatInterval = this.state.repeatInterval;
-
-        newEvent = {
-          userId: this.props.userId,
-          title: this.state.title,
-          rrule: {
-            freq: this.state.recurrencePeriod,
-            interval: this.state.repeatInterval,
-            dtstart: parsedRecurStart,
-            until: this.state.recurrenceEndDate,
-          },
-          duration: duration,
-          backgroundColor: "yellow",
-        };
-      } else if (this.state.recurrencePeriod === "weekly") {
-        newRepair.repeatInterval = this.state.repeatInterval;
-
-        newEvent = {
-          userId: this.props.userId,
-          title: this.state.title,
-          rrule: {
-            freq: this.state.recurrencePeriod,
-            interval: this.state.repeatInterval,
-            byweekday: this.state.repeatDayOfWeek,
-            dtstart: parsedRecurStart,
-            until: this.state.recurrenceEndDate,
-          },
-          duration: duration,
-          backgroundColor: "green",
-        };
-      }
-    } else {
-      let momentStart = this.state.startDate + " " + this.state.startTime;
-      let momentEnd = this.state.startDate + " " + this.state.endTime;
-      let duration = moment
-        .duration(
-          moment(momentEnd, "YYYY/MM/DD HH:mm").diff(
-            moment(momentStart, "YYYY/MM/DD HH:mm")
-          )
-        )
-        .asHours();
-
-      console.log("duration: " + duration);
-
-      let parsedStart =
-        this.state.startDate + "T" + this.state.startTime + ":00";
-      let parsedEnd = this.state.startDate + "T" + this.state.endTime + ":00";
-
-      newRepair.startDate = parsedStart;
-      newRepair.duration = duration;
-
-      newEvent = {
-        userId: this.props.userId,
-        category: this.state.repairType,
-        title: this.state.title,
-        start: parsedStart,
-        end: parsedEnd,
-        editable: true,
-      };
+    constructor(props) {
+      super(props);
     }
+    
+    render() {
+        return (
 
-    API.saveRepair(newRepair)
-      .then(newRepair => {
-        //        console.log("newRepair: " + JSON.stringify(newRepair.data._id));
-        return JSON.stringify(newRepair.data._id);
-      })
-      .then(repairId => {
-        newEvent.repairId = repairId;
-        API.saveEvent(newEvent);
-      })
-      .then(newEvent => {
-        //      console.log("newEvent in saveEvent: " + JSON.stringify(newEvent));
-        this.toggle();
-        this.loadEvents();
-      })
-      .catch(err => console.log(err));
-
-    this.setState({
-      repairType: "",
-      title: "",
-      startDate: "",
-      recurrencePeriod: "never",
-      repeatInterval: 1,
-      repeatDayOfWeek: "SU",
-      recurrenceStartDate: "",
-      recurrenceEndDate: "",
-      startTime: "",
-      endTime: "",
-      cost: 0,
-      priority: "low",
-      status: "Thinking about it!",
-      isVendor: false,
-      vendor: "",
-      notes: "",
-      editable: true,
-    });
-  };
-
-  render() {
-    return (
-      <div>
-        <div className="container">
-          <button
-            className="btn btn-info btn-lg mb-2"
-            onClick={this.handleNewEvent}
-          >
-            Add Event
-          </button>
-
-          <div className="calendar-container">
-            <div className="calendar-top" />
-            <div className="calendar" />
-            <FullCalendar
-              defaultView="dayGridMonth"
-              dateClick={this.handleDateClick}
-              /* this.handleDateClick */
-
-              header={{
-                left: "prev,next today",
-                center: "title",
-                right: "dayGridMonth,timeGridWeek,timeGridDay,listWeek",
-              }}
-              plugins={[dayGridPlugin, interactionPlugin, rrulePlugin]}
-              ref={this.calendarComponentRef}
-              events={this.state.events}
-            />
-          </div>
-          <div>
-            <Modal
-              isOpen={this.state.modal}
-              toggle={this.toggle}
-              className={this.props.className}
-            >
-              <Button className="ml-auto" color="danger" onClick={this.toggle}>
-                x
-              </Button>
-              <ModalBody>
-                <Form>
+              <Form>
                   <FormGroup>
                     <Label htmlFor="repairType">Type</Label>
                     <Input
@@ -556,11 +286,6 @@ export default class Calendar extends React.Component {
                     Add
                   </Button>
                 </Form>
-              </ModalBody>
-            </Modal>
-          </div>
-        </div>
-      </div>
-    );
-  }
+        )
+    }
 }
