+++ client/src/components/private/Repairs.js
@@ -10,6 +10,7 @@ import {
   Label,
   Input,
   Card,
+//  CardHeader,
   CardText,
   CardBody,
   CardTitle,
@@ -130,10 +131,8 @@ export default class Repair extends React.Component {
   };
 
   handleInputChange = event => {
-    // Getting the value and name of the input which triggered the change
     const { name, value } = event.target;
     console.log(event.target);
-    // Updating the input's state
     this.setState({
       [name]: value,
     });
@@ -152,9 +151,7 @@ export default class Repair extends React.Component {
       .catch(err => console.log(err));
   }
 
-  //// NOT COMPLETE ////
   handleFormSubmit = event => {
-    // Preventing the default behavior of the form submit (which is to refresh the page)
     event.preventDefault();
 
     let modifiedEvent = {};
@@ -190,21 +187,13 @@ export default class Repair extends React.Component {
 
       let parsedRecurStart =
         this.state.recurrenceStartDate + "T" + this.state.startTime + ":00";
-      //    let parsedRecurEnd = this.state.recurrenceEndDate + "T" + this.state.endTime + ":00";
-
+  
       modifiedRepair.recurrenceStartDate = parsedRecurStart;
       modifiedRepair.repeatDayOfWeek = this.state.repeatDayOfWeek;
       modifiedRepair.recurrenceEndDate = this.state.recurrenceEndDate;
       modifiedRepair.duration = duration;
 
       if (this.state.recurrencePeriod === "daily") {
-        /*     let momentDayInterval = moment
-                .duration(moment(this.state.recurrenceEndDate, 'YYYY/MM/DD')
-                .diff(moment(this.state.recurrenceStartDate, 'YYYY/MM/DD'))
-                ).asDays();
-            
-            console.log("momentDayInterval: " + momentDayInterval); */
-
         this.setState({
           repeatInterval: 1,
         });
@@ -213,7 +202,7 @@ export default class Repair extends React.Component {
 
         modifiedEvent = {
           userId: this.props.userId,
-          repairId: JSON.stringify(this.state.changeRepairId),
+          id: JSON.stringify(this.state.changeRepairId),
           title: this.state.title,
           rrule: {
             freq: this.state.recurrencePeriod,
@@ -235,7 +224,7 @@ export default class Repair extends React.Component {
             freq: this.state.recurrencePeriod,
             interval: this.state.repeatInterval,
             byweekday: this.state.repeatDayOfWeek,
-            dtstart: parsedRecurStart, // CHECK THIS //
+            dtstart: parsedRecurStart,
             until: this.state.recurrenceEndDate,
           },
           duration: duration,
@@ -264,7 +253,7 @@ export default class Repair extends React.Component {
 
       modifiedEvent = {
         userId: this.props.userId,
-        repairId: JSON.stringify(this.state.changeRepairId),
+        id: JSON.stringify(this.state.changeRepairId),
         category: this.state.repairType,
         title: this.state.title,
         start: parsedStart,
@@ -345,6 +334,7 @@ export default class Repair extends React.Component {
                     key={item.repairId}
                   >
                     <CardBody>
+                {/*     <CardHeader>{item.title}</CardHeader> */}
                       <CardTitle>
                         <h3>{item.title}</h3>
                       </CardTitle>
