+++ models/event.js
@@ -3,7 +3,7 @@ const Schema = mongoose.Schema;
 
 const eventSchema = new Schema({
   userId: { type: String },
-  repairId: { type: String },
+  id: { type: String },  // repairId
 //  group_id: { type: Schema.Types.ObjectId }
   title: { type: String },
   category: { type: String },
